complete moving into shared folder
