--- app/components/Breadcrumbs/BreadCrumbs.styles.tsx
@@ -1,5 +1,0 @@
-export const BreadCrumbsStyles = {
-    typography: {
-        color: '#190D03'
-    }
-};
