--- app/components/Link/Link.styles.ts
@@ -1,23 +1,0 @@
-
-export const linkStyles = {
-    button: {
-        cursor: 'pointer',
-        color: '#52545A',
-        display: 'flex',
-        borderRadius: 0,
-        '&:hover': {
-            backgroundColor: 'transparent',
-        },
-        '&:active': {
-            backgroundColor: 'transparent',
-        }
-    },
-    typography: {
-        display: 'flex',
-        alignItems: 'flex-end',
-        justifyContent: 'center',
-        '&:hover': {
-            borderBottom: '1px solid',
-        },
-    },
-};
