--- app/components/Link/NextLink.tsx
@@ -1,77 +1,0 @@
-'use client';
-import * as React from 'react';
-import clsx from 'clsx';
-import { useRouter } from 'next/router';
-import NextLink, { LinkProps as NextLinkProps } from 'next/link';
-import MuiLink, { LinkProps as MuiLinkProps } from '@mui/material/Link';
-
-interface NextLinkComposedProps
-  extends Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, 'href'>,
-    Omit<NextLinkProps, 'href' | 'as' | 'passHref' | 'onMouseEnter' | 'onClick' | 'onTouchStart'> {
-  to: NextLinkProps['href'];
-  linkAs?: NextLinkProps['as'];
-}
-
-export const NextLinkComposed = React.forwardRef<HTMLAnchorElement, NextLinkComposedProps>(
-  function NextLinkComposed(props, ref) {
-    const { to, linkAs, onClick, ...other } = props;
-
-    return <NextLink href={to} as={linkAs} ref={ref} onClick={onClick} {...other} />;
-  },
-);
-
-export type LinkProps = {
-  activeClassName?: string;
-  as?: NextLinkProps['as'];
-  href: NextLinkProps['href'];
-  linkAs?: NextLinkProps['as']; // Useful when the as prop is shallow by styled().
-  noLinkStyle?: boolean;
-} & Omit<NextLinkComposedProps, 'to' | 'linkAs' | 'href'> &
-  Omit<MuiLinkProps, 'href'>;
-
-// A styled version of the Next.js Link component:
-// https://nextjs.org/docs/pages/api-reference/components/link
-export const Link = React.forwardRef<HTMLAnchorElement, LinkProps>(function Link(props, ref) {
-  const {
-    activeClassName = 'active',
-    as,
-    className: classNameProps,
-    href,
-    linkAs: linkAsProp,
-    noLinkStyle,
-    ...other
-  } = props;
-
-  const router = useRouter();
-  const pathname = typeof href === 'string' ? href : href?.pathname;
-
-  const className = clsx(classNameProps, {
-    [activeClassName]: router.pathname === pathname && activeClassName,
-  });
-
-  const linkAs = linkAsProp ?? as ?? (href as string);
-  const nextjsProps = {
-    to: href,
-    linkAs,
-  };
-
-  if (noLinkStyle) {
-    return <NextLinkComposed className={className} ref={ref} {...nextjsProps} {...other} />;
-  }
-   const handleClick = (event: React.MouseEvent<HTMLAnchorElement>) => {
-    event.preventDefault();
-    router.push(href);
-  };
-  return (
-    <MuiLink
-    onClick={handleClick}
-      component={NextLinkComposed}
-      className={className}
-      ref={ref}
-      {...nextjsProps}
-      {...other}
-    />
-  );
-});
-
-export { NextLink };
