+++ app/shared/components/Breadcrumbs/BreadCrumbs.styles.tsx
@@ -1,0 +1,10 @@
+export const BreadCrumbsStyles = {
+  typography: {
+    color: '#190D03'
+  },
+  box: {
+    '&hover': {
+      textDecoration: 'underline'
+    }
+  }
+};
